refactor: change manyMediaInputs to MediaInput
